NLP-18 Reorganize scrubbing code
